requirements.txt
streamlit
requests
pandas
folium
python-dotenv